[feature] roi feat add
